+++ loop-fs-iso-impl/src/main/java/com/github/stephenc/javaisotools/loopfs/iso9660/Iso9660FileSystem.java
@@ -52,7 +52,7 @@ public class Iso9660FileSystem extends AbstractBlockFileSystem<Iso9660FileEntry>
         return buf;
     }
 
-    int readBytes(Iso9660FileEntry entry, int entryOffset, byte[] buffer, int bufferOffset, int len)
+    int readBytes(Iso9660FileEntry entry, long entryOffset, byte[] buffer, int bufferOffset, int len)
             throws IOException {
         long startPos = (entry.getStartBlock() * Constants.DEFAULT_BLOCK_SIZE) + entryOffset;
         return readData(startPos, buffer, bufferOffset, len);
